Adds ERC20 basic contract; Connect wallet

+++ app/contracts/SpadeRadioERC20Token.sol
@@ -2,22 +2,22 @@
 
 pragma solidity ^0.8.5;
 
-import 'https://github.com/OpenZeppelin/openzeppelin-contracts/blob/master/contracts/token/ERC20/ERC20.sol';
+import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract SpadeRadioERC20Token is ERC20 {
     address public admin;
-    
-    constructor() ERC20('Spade Radio', 'SPADA') {
-        _mint(msg.sender, 69420 * 10 ** 18);
+
+    constructor() ERC20("Spade Radio", "SPADA") {
+        _mint(msg.sender, 69420 * 10**18);
         admin = msg.sender;
-    }    
-    
-    function mint(address to, uint amount) external {
-        require(msg.sender == admin, 'Only Spada can mint $SPADE');
+    }
+
+    function mint(address to, uint256 amount) external {
+        require(msg.sender == admin, "Only Spada can mint $SPADE");
         _mint(to, amount);
     }
-    
-    function burn(uint amount) external {
+
+    function burn(uint256 amount) external {
         _burn(msg.sender, amount);
     }
 }
